homepage.html
Round 1 for the homepage. The script that is commented out at the very bottom is what pulls the Google Maps data into div #map. We'll need to figure out how to replace the current value for mtRecommend (right now it's just statically set to Copper Mountain's location) but that won't work til we have our decision algorithm built out.